Implements code review feedback
- Separates each scale heuristic in its own file;
- Adds half to even rounding for delay computation;
- Cleanup of comments and old code in `FabanBenchmarkGeneratorResource`.
